PushToTalk: Detect running apps on start
PushToTalk only changes state based on running applications if they are
launched or quit after PushToTalk has already been started. So if any
targeted apps are running when PushToTalk itself is started, or
Hammerspoon reloaded, it will not automatically set the microphone
state.

This add a disabled by default option to check the list of running
applications when starting PushToTalk, so relaunching or reloading
Hammerspoon will typically get you back to the desired state, instead
of default state regardless of what apps are running or not.

The `initialState()` function is optimistic, in that it will use the
configured state for whichever application from `app_switcher` it
encounters first, so for complex configurations, a relaunch/reload of
Hammerspoon might not set the microphone state back to what it
originally was.