Added `eject_on_lid_close` parameter
The detection is done based on an hs.screen.watcher which detects when
the internal display disappears.

This is useful when connected to an external display, because closing
the lid does not put the system to sleep, but might still be the
prelude to disconnecting the machine from the display.