ClipboardTool: fixes image support
There were some overlooked references to the earlier content of a copied
item (which could only be a string), so these were changed to the value
of the newly implemented table that also defines whether it's regular
text or an image.

Also updated the chooser to show bigger text and preview image date for
copied images.