Arrange Desktop 2.0 (#247)
* Upgrade ArrangeDesktop to 2.0

This overhauls the functionality of the Spoon to auto-generate the configuration of your desktop arrangement. This replaces the previous flow of having to select a menu option, open the console, hand copy the config from there into your `init.lua`, and tweak to get it running.

* Upgrade ArrangeDesktop to 2.0

Regenerate documentation JSON.

* Upgrade ArrangeDesktop to 2.0

Remove reference to blog post in docs.