Updating for Asana 1.1 and adding body
The body of the POST can be used to provide all the necessary data for
an Asana task. In order for the server to read the body the content
ecooding must be declared and provided as a header. This allows for
unlimited Asana templates defined in Hammerspoon. I have not provided an
example.