fix seal plugins cmd being ignored in commands table
Seal plugins can define a table of commands which contains
the command used to launch it in the `cmd` field.

However currently the key of the command in the commands table
is used instead of this field.
This was identical for most seal plugin except filesearch which
uses the special ' character as the command trigger.