Fix query replacement for non-spaced queries
`string.gsub` returns two values: the string with substitutions and the number of substitutions made

If doing the nested `gsub`, these two return values will be passed into the outer `gsub`, which accepts an optional fourth parameter specifying how many substitutions should be made. So the number of substitutions in the inner `gsub` determines the number of substitutions in the outer `gsub`: When the query doesn't contain a blank, that number of substitutions in the inner `gsub` is `0` and the outer `gsub` won't replace anything, leaving '${query}' in the `url`.