Allow reading URL patterns from a file
If the first element of a url_patterns element points to a file, the
list of patterns to match is read from it. Furthermore, a watcher is set
so that the patterns are automatically reloaded if the file changes.

This avoids having to include URLs in the main config, and also having
to restart Hammerspoon whenever a pattern is added or changes.