Improve MicMute spoon by optionally displaying an icon.
This patch adds the possibility to display an icon in the menubar.
The default behaviour of the spoon is not changed. The icon display
can be activated with the new configuration `display_mode`.

Also, the menubar text can be customised by the user, with the new
configurations `title_on` and `title_off`.