Fix a race condition issue in Seal filesearch plugin
Updated the Seal filesearch plugin to resolve a race
condition which occurred when a new spotlight search
was initiated before the completion of a previous search.
This issue resulted in occasional receipt of obsolete
`didFinish` spotlight events.

To mitigate this, a unique `hs.spotlight` object is now
assigned to each new search, and we disregard any event
or callback that is not related to the current search.

Additionally, the spotlight file search logic has been
isolated into a distinct class for improved clarity.