Improve responsiveness of Seal filesearch plugin
Contrary to most plugins, the Seal filesearch plugin
doesn't quickly filter results from an existing set
in memory, but it triggers an actual spotlight search
each time.

That is more resource-intensive and calling it too
frequently for each key typed does hurt the responsiveness
of the Seal chooser (lua being single-threaded).

To avoid that, the plugin now uses it own `queryChangedTimerDuration`
parameter so it can be set to higher value compared to the
defaut Seal one.
The default of 100mn provides a good responsiveness
while still display results with an acceptable delay.