Improve search results order in Seal filesearch plugin
Spotlight results naturally put first the items that
were recently accessed.
However it doesn't put first the item that really
contain the word as it (not as a sub-part of the word)
whereas it is often the ones you want first typically.

We emulate that behaviour by re-ordering the results
to boost the item that contains whole words.