MouseFollowsFocus: avoid getting the mouse stuck in screen corners (#335)
The mouse cursor moves in a straight line between its current position
and the destination set with hs.mouse.absolutePosition(). If the mouse
is moving to a new screen, it can get caught on a corner that does not
connect between screens, depending on the screen sizes and layout.

When moving the mouse between screens, first move the mouse to the
center of the current screen, then to the center of the new screen, then
the target destination. This path has the best chance at avoiding any
non-connected edges between screens.

These extra mouse movements happen instantaneously and are not visible
to the user.